backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
websockets>=12.0
python-multipart>=0.0.9

# LangGraph / LangChain
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.3.0

# Terminal
ptyprocess>=0.7.0      # Unix PTY
pywinpty>=2.0.0        # Windows PTY

# Config
pydantic>=2.7.0
pydantic-settings>=2.3.0
pyyaml>=6.0

# HTTP client (for Prometheus / Loki / MiniMax)
httpx>=0.27.0
requests>=2.32.0
